WEB-3177: rename Campaign Manager references to Campaign Assistant
