+++ app/(candidate)/dashboard/campaign-assistant/components/DeleteThread.js
