+++ app/(candidate)/dashboard/campaign-assistant/components/ajaxActions.js
