+++ app/(candidate)/dashboard/campaign-assistant/page.js
@@ -1,13 +1,13 @@
 import { fetchUserCampaign } from 'app/(candidate)/onboarding/shared/getCampaign';
 import pageMetaData from 'helpers/metadataHelper';
-import CampaignManagerPage from './components/CampaignManagerPage';
+import CampaignAssistantPage from './components/CampaignAssistantPage';
 import { getServerUser } from 'helpers/userServerHelper';
 import { adminAccessOnly } from 'helpers/permissionHelper';
 
 const meta = pageMetaData({
-  title: 'Campaign Manager | GoodParty.org',
-  description: 'Campaign Manager',
-  slug: '/dashboard/campaign-manager',
+  title: 'Campaign Assistant | GoodParty.org',
+  description: 'Campaign Assistant',
+  slug: '/dashboard/campaign-assistant',
 });
 export const metadata = meta;
 
@@ -19,10 +19,10 @@ export default async function Page({ params, searchParams }) {
   const { campaign } = await fetchUserCampaign();
 
   const childProps = {
-    pathname: '/dashboard/campaign-manager',
+    pathname: '/dashboard/campaign-assistant',
     user,
     campaign,
   };
 
-  return <CampaignManagerPage {...childProps} />;
+  return <CampaignAssistantPage {...childProps} />;
 }
